+++ flask-server/static/script/play/amazons.js
@@ -131,10 +131,18 @@ function update_cell_image(cell, type, add_or_remove)
         class_name = "cell-fire"
     }
 
+    el_clist = document.getElementById("cell-" + cell).classList;
+
     if(add_or_remove == 0){
-        document.getElementById("cell-" + cell).classList.add(class_name)
+        el_clist.add(class_name)
+        if(class_name.includes("queen")){
+            el_clist.add("cell-queen")
+        }
     }else{
-        document.getElementById("cell-" + cell).classList.remove(class_name)
+        el_clist.remove(class_name)
+        if(class_name.includes("queen")){
+            el_clist.remove("cell-queen")
+        }
     }
 }
 
@@ -189,24 +197,31 @@ function cell_can_reach(from, to){
     return flag
 }
 
-function valid_move_indicator(cell, mode){
-    if(mode == 1){
+function valid_move_indicator(cell, removeOrAdd, moveOrShoot){
+
+    if(moveOrShoot == "move"){
+        indicatorName = "valid-move-indicator"
+    }else{
+        indicatorName = "valid-shoot-indicator"
+    }
+
+    if(removeOrAdd === "remove"){ // remove indicators
         for (let x = 0; x < game_size; x++) {
             for (let y = 0; y < game_size; y++) {
-                if(document.getElementById("cell-" + x.toString() + y.toString()).classList.contains("valid-move-indicator")){
-                    document.getElementById("cell-" + x.toString() + y.toString()).classList.remove("valid-move-indicator")
+                cellClasses = document.getElementById("cell-" + x.toString() + y.toString()).classList;
+                if(cellClasses.contains(indicatorName)){
+                    cellClasses.remove(indicatorName)
                 }
             }
         }
-    }else {
-
+    }else{ // add indicators
         c_x = Number(cell[0])
         c_y = Number(cell[1])
         for (let x = 0; x < game_size; x++) {
             for (let y = 0; y < game_size; y++) {
                 if(cell_can_reach(cell,x.toString() + y.toString())){
                     
-                    document.getElementById("cell-" + x.toString() + y.toString()).classList.add("valid-move-indicator")
+                    document.getElementById("cell-" + x.toString() + y.toString()).classList.add(indicatorName)
                     
                 }
             }
@@ -228,11 +243,11 @@ function reset_move(msg, cont){
 
     if(temp_source){
         document.getElementById("cell-" + temp_source).classList.remove("selected-indicator-source");
-        valid_move_indicator(temp_source,1)
+        valid_move_indicator(temp_source,"remove","move")
     }
     if(temp_move){
         document.getElementById("cell-" + temp_move).classList.remove("selected-indicator-move");
-        valid_move_indicator(temp_move,1,temp_source)
+        valid_move_indicator(temp_move,"remove","shoot")
     }
     temp_move = null 
     temp_source = null
@@ -251,7 +266,7 @@ function handle_click(cell){
             temp_source = cell
             board[Number(temp_source.charAt(0))][Number(temp_source.charAt(1))] = 0
             document.getElementById("cell-" + cell).classList.add("selected-indicator-source");
-            valid_move_indicator(temp_source,0)
+            valid_move_indicator(temp_source,"add","move")
             
             temp_step += 1;
         }else{
@@ -262,8 +277,8 @@ function handle_click(cell){
         if(cell_can_reach(temp_source,cell)){
             temp_move = cell
             board[Number(temp_move.charAt(0))][Number(temp_move.charAt(1))] = turn + 1
-            valid_move_indicator(temp_source,1)
-            valid_move_indicator(temp_move,0,temp_source)
+            valid_move_indicator(temp_source,"remove","move")
+            valid_move_indicator(temp_move,"add", "shoot")
             document.getElementById("cell-" + cell).classList.add("selected-indicator-move");
             temp_step += 1;
         }else{
@@ -272,7 +287,7 @@ function handle_click(cell){
     }else if(temp_step == 2){
         temp_shoot = cell
         if(cell_can_reach(temp_move,temp_shoot,temp_source)){
-            valid_move_indicator(temp_move,1)
+            valid_move_indicator(temp_move,"remove", "shoot")
             board[Number(temp_shoot.charAt(0))][Number(temp_shoot.charAt(1))] = 3
             play(temp_source, temp_move, temp_shoot);
         }else{