+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 - Establish simple room-joining logic.
 - Emit relevant updates when joining rooms.
 - Sync moves, resulting in basic play!
+- Added back-end move validation
+- Amazons move indicator updates
+    - Better JS code signatures
+    - Show different indicator on queen
+    - Show different colors for move and shoot
 # 2021-10-01
 - Restarting this project with a focus on the ability to play correspondence games against other players.
 - Cleared up instructions for running the codebase.